chore(atomix): prepare 0.24 for rolling upgrade
