+++ atomix/cluster/src/main/java/io/atomix/cluster/messaging/impl/DefaultClusterEventService.java
@@ -32,7 +32,8 @@ import io.atomix.cluster.messaging.ManagedClusterEventService;
 import io.atomix.cluster.messaging.MessagingService;
 import io.atomix.cluster.messaging.Subscription;
 import io.atomix.utils.net.Address;
-import io.atomix.utils.serializer.Namespace;
+import io.atomix.utils.serializer.FallbackNamespace;
+import io.atomix.utils.serializer.NamespaceImpl;
 import io.atomix.utils.serializer.Namespaces;
 import io.atomix.utils.serializer.Serializer;
 import io.atomix.utils.time.LogicalTimestamp;
@@ -63,12 +64,12 @@ public class DefaultClusterEventService
 
   private static final Serializer SERIALIZER =
       Serializer.using(
-          Namespace.builder()
-              .register(Namespaces.BASIC)
-              .register(MemberId.class)
-              .register(LogicalTimestamp.class)
-              .register(WallClockTimestamp.class)
-              .build());
+          new FallbackNamespace(
+              new NamespaceImpl.Builder()
+                  .register(Namespaces.BASIC)
+                  .register(MemberId.class)
+                  .register(LogicalTimestamp.class)
+                  .register(WallClockTimestamp.class)));
 
   private static final String SUBSCRIPTION_PROPERTY_NAME = "event-service-topics-subscribed";
   private final ClusterMembershipService membershipService;
