+++ atomix/cluster/src/main/java/io/atomix/cluster/messaging/impl/NettyUnicastService.java
@@ -25,7 +25,8 @@ import io.atomix.cluster.messaging.ManagedUnicastService;
 import io.atomix.cluster.messaging.MessagingConfig;
 import io.atomix.cluster.messaging.UnicastService;
 import io.atomix.utils.net.Address;
-import io.atomix.utils.serializer.Namespace;
+import io.atomix.utils.serializer.FallbackNamespace;
+import io.atomix.utils.serializer.NamespaceImpl.Builder;
 import io.atomix.utils.serializer.Namespaces;
 import io.atomix.utils.serializer.Serializer;
 import io.netty.bootstrap.Bootstrap;
@@ -56,12 +57,12 @@ public class NettyUnicastService implements ManagedUnicastService {
   private static final Logger LOGGER = LoggerFactory.getLogger(NettyUnicastService.class);
   private static final Serializer SERIALIZER =
       Serializer.using(
-          Namespace.builder()
-              .register(Namespaces.BASIC)
-              .nextId(Namespaces.BEGIN_USER_CUSTOM_ID)
-              .register(Message.class)
-              .register(new AddressSerializer(), Address.class)
-              .build());
+          new FallbackNamespace(
+              new Builder()
+                  .register(Namespaces.BASIC)
+                  .nextId(Namespaces.BEGIN_USER_CUSTOM_ID)
+                  .register(Message.class)
+                  .register(new AddressSerializer(), Address.class)));
 
   private final Logger log = LoggerFactory.getLogger(getClass());
