+++ atomix/cluster/src/main/java/io/atomix/raft/partition/RaftPartitionGroup.java
@@ -42,7 +42,9 @@ import io.atomix.utils.concurrent.ThreadContextFactory;
 import io.atomix.utils.logging.ContextualLoggerFactory;
 import io.atomix.utils.logging.LoggerContext;
 import io.atomix.utils.memory.MemorySize;
+import io.atomix.utils.serializer.FallbackNamespace;
 import io.atomix.utils.serializer.Namespace;
+import io.atomix.utils.serializer.NamespaceImpl;
 import io.atomix.utils.serializer.Namespaces;
 import java.io.File;
 import java.time.Duration;
@@ -278,13 +280,13 @@ public class RaftPartitionGroup implements ManagedPartitionGroup {
 
     @Override
     public Namespace namespace() {
-      return Namespace.builder()
-          .nextId(Namespaces.BEGIN_USER_CUSTOM_ID + 100)
-          .register(RaftPartitionGroupConfig.class)
-          .register(RaftStorageConfig.class)
-          .register(RaftCompactionConfig.class)
-          .register(StorageLevel.class)
-          .build();
+      return new FallbackNamespace(
+          new NamespaceImpl.Builder()
+              .nextId(Namespaces.BEGIN_USER_CUSTOM_ID + 100)
+              .register(RaftPartitionGroupConfig.class)
+              .register(RaftStorageConfig.class)
+              .register(RaftCompactionConfig.class)
+              .register(StorageLevel.class));
     }
 
     @Override
