+++ atomix/cluster/src/main/java/io/atomix/raft/partition/impl/RaftNamespaces.java
@@ -41,7 +41,10 @@ import io.atomix.raft.storage.log.entry.ConfigurationEntry;
 import io.atomix.raft.storage.log.entry.InitializeEntry;
 import io.atomix.raft.storage.system.Configuration;
 import io.atomix.raft.zeebe.ZeebeEntry;
+import io.atomix.utils.serializer.FallbackNamespace;
 import io.atomix.utils.serializer.Namespace;
+import io.atomix.utils.serializer.NamespaceImpl;
+import io.atomix.utils.serializer.NamespaceImpl.Builder;
 import io.atomix.utils.serializer.Namespaces;
 import java.time.Instant;
 import java.util.ArrayList;
@@ -54,72 +57,7 @@ public final class RaftNamespaces {
 
   /** Raft protocol namespace. */
   public static final Namespace RAFT_PROTOCOL =
-      Namespace.builder()
-          .register(Namespaces.BASIC)
-          .nextId(Namespaces.BEGIN_USER_CUSTOM_ID)
-          .register(Void.class) // OpenSessionRequest
-          .register(Void.class) // OpenSessionResponse
-          .register(Void.class) // CloseSessionRequest
-          .register(Void.class) // CloseSessionResponse
-          .register(Void.class) // KeepAliveRequest
-          .register(Void.class) // KeepAliveResponse
-          .register(Void.class) // HeartbeatRequest
-          .register(Void.class) // HeartbeatResponse
-          .register(Void.class) // QueryRequest
-          .register(Void.class) // QueryResponse
-          .register(Void.class) // CommandRequest
-          .register(Void.class) // CommandResponse
-          .register(Void.class) // MetadataRequest
-          .register(Void.class) // MetadataResponse
-          .register(JoinRequest.class)
-          .register(JoinResponse.class)
-          .register(LeaveRequest.class)
-          .register(LeaveResponse.class)
-          .register(ConfigureRequest.class)
-          .register(ConfigureResponse.class)
-          .register(ReconfigureRequest.class)
-          .register(ReconfigureResponse.class)
-          .register(InstallRequest.class)
-          .register(InstallResponse.class)
-          .register(PollRequest.class)
-          .register(PollResponse.class)
-          .register(VoteRequest.class)
-          .register(VoteResponse.class)
-          .register(AppendRequest.class)
-          .register(AppendResponse.class)
-          .register(Void.class) // PublishRequest
-          .register(Void.class) // ResetRequest
-          .register(RaftResponse.Status.class)
-          .register(RaftError.class)
-          .register(RaftError.Type.class)
-          .register(Void.class) // ReadConsistency
-          .register(Void.class) // SessionMetadata
-          .register(Void.class) // CloseSessionEntry
-          .register(Void.class) // CommandEntry
-          .register(ConfigurationEntry.class)
-          .register(InitializeEntry.class)
-          .register(Void.class) // KeepAliveEntry
-          .register(Void.class) // MetadataEntry
-          .register(Void.class) // OpenSessionEntry
-          .register(Void.class) // QueryEntry
-          .register(Void.class) // PrimitiveOperation
-          .register(Void.class) // PrimitiveEvent
-          .register(Void.class) // DefaultEventType
-          .register(Void.class) // DefaultOperationId
-          .register(Void.class) // OperationType
-          .register(Void.class) // ReadConsistency
-          .register(ArrayList.class)
-          .register(LinkedList.class)
-          .register(Collections.emptyList().getClass())
-          .register(HashSet.class)
-          .register(DefaultRaftMember.class)
-          .register(MemberId.class)
-          .register(Void.class) // SessionId
-          .register(RaftMember.Type.class)
-          .register(Instant.class)
-          .register(Configuration.class)
-          .register(ZeebeEntry.class)
-          .build("RaftProtocol");
+      new FallbackNamespace(buildProtocolNamespace().name("RaftProtocol"));
 
   /**
    * Raft storage namespace.
@@ -127,31 +65,102 @@ public final class RaftNamespaces {
    * <p>*Be aware* we use the Void type for replaced/removed types to keep the id's of used types,
    * otherwise we break compatibility.
    */
-  public static final Namespace RAFT_STORAGE =
-      Namespace.builder()
-          .register(Namespaces.BASIC)
-          .nextId(Namespaces.BEGIN_USER_CUSTOM_ID + 100)
-          .register(Void.class) // CloseSessionEntry
-          .register(Void.class) // CommandEntry
-          .register(ConfigurationEntry.class)
-          .register(InitializeEntry.class)
-          .register(Void.class) // KeepAliveEntry
-          .register(Void.class) // MetadataEntry
-          .register(Void.class) // OpenSessionEntry
-          .register(Void.class) // QueryEntry
-          .register(Void.class) // PrimitiveOperation
-          .register(Void.class) // DefaultOperationId
-          .register(Void.class) // OperationType
-          .register(Void.class) // ReadConsistency
-          .register(ArrayList.class)
-          .register(HashSet.class)
-          .register(DefaultRaftMember.class)
-          .register(MemberId.class)
-          .register(RaftMember.Type.class)
-          .register(Instant.class)
-          .register(Configuration.class)
-          .register(ZeebeEntry.class)
-          .build("RaftStorage");
+  public static final FallbackNamespace RAFT_STORAGE =
+      new FallbackNamespace(buildStorageNamespace().name("RaftStorage"));
 
   private RaftNamespaces() {}
+
+  private static Builder buildStorageNamespace() {
+    return new NamespaceImpl.Builder()
+        .register(Namespaces.BASIC)
+        .nextId(Namespaces.BEGIN_USER_CUSTOM_ID + 100)
+        .register(Void.class) // CloseSessionEntry
+        .register(Void.class) // CommandEntry
+        .register(ConfigurationEntry.class)
+        .register(InitializeEntry.class)
+        .register(Void.class) // KeepAliveEntry
+        .register(Void.class) // MetadataEntry
+        .register(Void.class) // OpenSessionEntry
+        .register(Void.class) // QueryEntry
+        .register(Void.class) // PrimitiveOperation
+        .register(Void.class) // DefaultOperationId
+        .register(Void.class) // OperationType
+        .register(Void.class) // ReadConsistency
+        .register(ArrayList.class)
+        .register(HashSet.class)
+        .register(DefaultRaftMember.class)
+        .register(MemberId.class)
+        .register(RaftMember.Type.class)
+        .register(Instant.class)
+        .register(Configuration.class)
+        .register(ZeebeEntry.class);
+  }
+
+  private static Builder buildProtocolNamespace() {
+    return new NamespaceImpl.Builder()
+        .register(Namespaces.BASIC)
+        .nextId(Namespaces.BEGIN_USER_CUSTOM_ID)
+        .register(Void.class) // OpenSessionRequest
+        .register(Void.class) // OpenSessionResponse
+        .register(Void.class) // CloseSessionRequest
+        .register(Void.class) // CloseSessionResponse
+        .register(Void.class) // KeepAliveRequest
+        .register(Void.class) // KeepAliveResponse
+        .register(Void.class) // HeartbeatRequest
+        .register(Void.class) // HeartbeatResponse
+        .register(Void.class) // QueryRequest
+        .register(Void.class) // QueryResponse
+        .register(Void.class) // CommandRequest
+        .register(Void.class) // CommandResponse
+        .register(Void.class) // MetadataRequest
+        .register(Void.class) // MetadataResponse
+        .register(JoinRequest.class)
+        .register(JoinResponse.class)
+        .register(LeaveRequest.class)
+        .register(LeaveResponse.class)
+        .register(ConfigureRequest.class)
+        .register(ConfigureResponse.class)
+        .register(ReconfigureRequest.class)
+        .register(ReconfigureResponse.class)
+        .register(InstallRequest.class)
+        .register(InstallResponse.class)
+        .register(PollRequest.class)
+        .register(PollResponse.class)
+        .register(VoteRequest.class)
+        .register(VoteResponse.class)
+        .register(AppendRequest.class)
+        .register(AppendResponse.class)
+        .register(Void.class) // PublishRequest
+        .register(Void.class) // ResetRequest
+        .register(RaftResponse.Status.class)
+        .register(RaftError.class)
+        .register(RaftError.Type.class)
+        .register(Void.class) // ReadConsistency
+        .register(Void.class) // SessionMetadata
+        .register(Void.class) // CloseSessionEntry
+        .register(Void.class) // CommandEntry
+        .register(ConfigurationEntry.class)
+        .register(InitializeEntry.class)
+        .register(Void.class) // KeepAliveEntry
+        .register(Void.class) // MetadataEntry
+        .register(Void.class) // OpenSessionEntry
+        .register(Void.class) // QueryEntry
+        .register(Void.class) // PrimitiveOperation
+        .register(Void.class) // PrimitiveEvent
+        .register(Void.class) // DefaultEventType
+        .register(Void.class) // DefaultOperationId
+        .register(Void.class) // OperationType
+        .register(Void.class) // ReadConsistency
+        .register(ArrayList.class)
+        .register(LinkedList.class)
+        .register(Collections.emptyList().getClass())
+        .register(HashSet.class)
+        .register(DefaultRaftMember.class)
+        .register(MemberId.class)
+        .register(Void.class) // SessionId
+        .register(RaftMember.Type.class)
+        .register(Instant.class)
+        .register(Configuration.class)
+        .register(ZeebeEntry.class);
+  }
 }
