+++ atomix/storage/src/test/java/io/atomix/storage/journal/FileChannelJournalSegmentReaderTest.java
@@ -22,6 +22,7 @@ import io.atomix.storage.StorageLevel;
 import io.atomix.storage.journal.JournalReader.Mode;
 import io.atomix.storage.journal.index.SparseJournalIndex;
 import io.atomix.utils.serializer.Namespace;
+import io.atomix.utils.serializer.NamespaceImpl;
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -33,7 +34,8 @@ import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 
 public class FileChannelJournalSegmentReaderTest {
-  private static final Namespace NAMESPACE = Namespace.builder().register(Integer.class).build();
+  private static final Namespace NAMESPACE =
+      new NamespaceImpl.Builder().register(Integer.class).build();
   private static final Integer ENTRY = 1;
   private static final int ENTRY_SIZE =
       NAMESPACE.serialize(ENTRY).length + Integer.BYTES; // padding for checksum;
